refactor(core): stop exporting NIP-01 types from client.ts or relay.ts
