+++ core/nodes.ts
@@ -1,4 +1,5 @@
-import type { Logger, NostrMessage } from "./types.ts";
+import type { NostrMessage } from "./nips/01.ts";
+import type { Logger } from "./types.ts";
 import { WebSocketLike, WebSocketReadyState } from "./websockets.ts";
 import { NonExclusiveWritableStream } from "./streams.ts";
